Improved Deconstructors, refactored Print2D()

+++ PointerAndReferencePractice/vector.h
@@ -9,16 +9,16 @@ class Vector2D : public Point2D
 	private:
 		
 		double* vectorDir = new double[2]; // Direction of the Vector, clamped between 0 and 1 along both Axis
-		int vectorMagnitude = 0;
+		double vectorMagnitude = 0;
 
 
 	public:
 
-		Vector2D(const std::string& _name, int _magnitude, const std::vector<double>& _vecDir, int posX, int posY); // Constuctor
+		Vector2D(const std::string& _name, double _magnitude, const std::vector<double>& _vecDir, int posX, int posY); // Constuctor
 		~Vector2D() override; // Deconstructor
 
 		// Getters
-		int GetVectorMagnitude() const;
+		double GetVectorMagnitude() const;
 		double* GetVectorDirection() const;
 
 		// Setters
